requirements.txt
Click==7.1.1
Flask==1.1.1
future==0.18.2
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pyserial==3.4
PyYAML==5.3.1
Werkzeug==1.0.0
